requirements.txt
PyGithub==1.*,>=1.55
black==21.*,>=21.12b0
flake8==4.*,>=4.0.1
pylint==2.*,>=2.12.2
python-lsp-server==1.*,>=1.3.3
jinja2==3.*,>=3.0.3
requests==2.*,>=2.27.1
types-requests==2.*,>=2.27.7
